Change uvIndex property type from Int to Double in CurrentWeatherDto

+++ app/src/main/java/com/pmalaquias/weatherforecast/data/remote/dto/CurrentWeatherDto.kt
@@ -22,7 +22,7 @@ data class CurrentWeatherDto(
     @SerializedName ("wind_dir") val windDir: String,
     @SerializedName("wind_kph") val windKph: Double,
     @SerializedName("feelslike_c") val feelslikeCelcius: Double,
-    @SerializedName("uv") val uvIndex: Int,
+    @SerializedName("uv") val uvIndex: Double,
     @SerializedName("is_day") val isDay: Int, // 1 for day, 0 for night
     @SerializedName("pressure_mb") val pressureMb: Double,
     @SerializedName("precip_mm") val precipitationMm: Double
